+++ docs/CHANGELOG.rst
@@ -22,6 +22,7 @@ Changed
 - **BACKWARD INCOMPATIBLE:** Drop support for ``Python 3.6``, ``Python 3.7``, 
   ``Python 3.8`` and ``Python 3.9``. The processing container continues to
   support ``Python >= 3.6``.
+- Add ``source`` attribute to observer notifications
 
 
 ===================